backend/api_server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
Pillow==11.1.0
pytesseract==0.3.13
pypdf==5.3.0
